+++ src/main/java/com/ssafyss/board_practice/todo/infrastructure/TodoRepository.java
@@ -1,11 +1,10 @@
 package com.ssafyss.board_practice.todo.infrastructure;
 
-import com.ssafyss.board_practice.todo.entity.Todo;
-import java.util.List;
+import com.ssafyss.board_practice.todo.domain.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface TodoRepository extends JpaRepository<Todo, Long> {
-    List<Todo> findByUserId(Long userId);
+
 }
